requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
requests==2.22.0
slackclient==2.0.1
urllib3==1.25.3
yarl==1.3.0
